% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_a_name.R
\name{draw_a_name}
\alias{draw_a_name}
\title{Title}
\usage{
draw_a_name(the_name, the_sex)
}
\arguments{
\item{the_name}{name that we are looking for}

\item{the_sex}{the sex of our guy}
}
\value{
a ggplot
}
\description{
Title
}
\examples{
\dontrun{
draw_a_name(the_name = "Vincent",the_sex = "M")
draw_a_name( "Diane", "F")
}
}
