% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_names.R
\name{draw_names}
\alias{draw_names}
\title{Draw the evolution af several names}
\usage{
draw_names(names)
}
\arguments{
\item{names}{the names you want to observe}
}
\value{
a ggplot
}
\description{
Draw the evolution af several names
}
\examples{
\dontrun{ draw_names(c("Vincent", "Alex"))}
}
