% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_csv_reader.R
\name{my_csv_reader}
\alias{my_csv_reader}
\title{Import multiple csv files}
\usage{
my_csv_reader(folder)
}
\arguments{
\item{folder}{path to the folder}
}
\value{
a list
}
\description{
Import multiple csv files
}
\examples{
\dontrun{
my_csv_reader("../../path/to/folder")
}
}
